testing

+++ test_ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: bokim <bokim@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/18 15:20:22 by bokim             #+#    #+#             */
-/*   Updated: 2026/02/18 15:23:25 by bokim            ###   ########.fr       */
+/*   Updated: 2026/02/18 15:24:22 by bokim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@
 int main(void)
 {
     char str[5] = "test";
-    int expected = strlen(str);
-    int actual = ft_strlen(str);
+    size_t expected = strlen(str);
+    size_t actual = ft_strlen(str);
     if (expected == actual)
         return (0);
     else
